Added support for CSLUT-based cell constraints. When combined with FOM beam constraints, we parse at ~100 ms/sentence on my Mac.
Changed cell storage order from horizontal to diagonal (e.g., cell 0,1; 0,2, 0,3 ... 0,2; 1,2,...). In theory, this should avoid a few cache misses when iterating over entries in child cells (e.g., when populating cell 0,4, we access 0,1; 0,2; 0,3 in that order). I arbitrarily chose to order on the left diagonal. I don't think we can do much to avoid the cache misses on one side or the other, although I might try to model that sometime. Seems to help just a bit. Only a couple percent, but we'll take it.

Eliminated a number of obsolete cartesian product functions

Hard-coded CscSpmvParser to use the segmented perfect hash. Hard-coding it allows us to compute part of the hash function once for each left child, instead of separately for each child pair. If the JIT were _really_ smart, it might manage to unroll some of those loops and inline the computation, but that would be really impressive, and this change also seems to help a bit, so I don't think the JIT's really managing that.

Stored first and last valid right sibling for each left child in the grammar, and skip x-product loops for impossible siblings.

+++ java/src/edu/ohsu/cslu/grammar/CsrSparseMatrixGrammar.java
@@ -77,7 +77,7 @@ public class CsrSparseMatrixGrammar extends SparseMatrixGrammar {
         }
 
         for (final Production p : binaryProductions) {
-            maps1[p.parent].put(cartesianProductFunction.pack(p.leftChild, p.rightChild), p.prob);
+            maps1[p.parent].put(cartesianProductFunction.pack((short) p.leftChild, (short) p.rightChild), p.prob);
         }
         final Int2FloatOpenHashMap[] maps = maps1;
