+++ java/src/edu/ohsu/cslu/parser/ml/CartesianProductHashSpmlParser.java
@@ -70,7 +70,7 @@ public class CartesianProductHashSpmlParser extends
             final int rightEnd = chart.maxRightChildIndex(rightCellIndex);
 
             for (int i = leftStart; i <= leftEnd; i++) {
-                final int leftChild = chart.nonTerminalIndices[i];
+                final short leftChild = chart.nonTerminalIndices[i];
 
                 // TODO Skip non-terminals which never occur as left children
